# Optimize CI/CD pipeline performance