--- a/.github/lighthouse/lighthouserc.js
+++ b/.github/lighthouse/lighthouserc.js
@@ -1,0 +1,31 @@
+module.exports = {
+  ci: {
+    collect: {
+      startServerCommand: 'npm run start',
+      startServerReadyPattern: 'ready on',
+      startServerReadyTimeout: 30000,
+      numberOfRuns: 3,
+      settings: {
+        preset: 'desktop',
+        skipAudits: ['uses-http2', 'uses-long-cache-ttl', 'uses-text-compression'],
+      },
+    },
+    assert: {
+      assertions: {
+        'categories:performance': ['warn', { minScore: 0.7 }],
+        'categories:accessibility': ['warn', { minScore: 0.8 }],
+        'categories:best-practices': ['warn', { minScore: 0.8 }],
+        'categories:seo': ['warn', { minScore: 0.8 }],
+        'first-contentful-paint': ['warn', { maxNumericValue: 2000 }],
+        'interactive': ['warn', { maxNumericValue: 5000 }],
+        'speed-index': ['warn', { maxNumericValue: 4000 }],
+        'total-blocking-time': ['warn', { maxNumericValue: 300 }],
+        'largest-contentful-paint': ['warn', { maxNumericValue: 3000 }],
+        'cumulative-layout-shift': ['warn', { maxNumericValue: 0.1 }],
+      },
+    },
+    upload: {
+      target: 'temporary-public-storage',
+    },
+  },
+};