--- a/.github/workflows/performance.yml
+++ b/.github/workflows/performance.yml
@@ -1,0 +1,130 @@
+name: Performance Monitoring
+
+on:
+  pull_request:
+    branches: [main]
+    paths:
+      - 'frontend/**'
+      - 'package*.json'
+
+jobs:
+  lighthouse:
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v4
+    
+    - name: Setup Node.js
+      uses: actions/setup-node@v4
+      with:
+        node-version: '20'
+        cache: 'npm'
+        cache-dependency-path: frontend/package-lock.json
+    
+    - name: Install dependencies
+      working-directory: ./frontend
+      run: npm ci
+    
+    - name: Build application
+      working-directory: ./frontend
+      run: npm run build
+    
+    - name: Run Lighthouse CI
+      uses: treosh/lighthouse-ci-action@v12
+      with:
+        urls: |
+          http://localhost:3000
+        uploadArtifacts: true
+        temporaryPublicStorage: true
+        runs: 3
+        configPath: './.github/lighthouse/lighthouserc.js'
+      continue-on-error: true
+
+  bundle-size:
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v4
+    
+    - name: Setup Node.js
+      uses: actions/setup-node@v4
+      with:
+        node-version: '20'
+        cache: 'npm'
+        cache-dependency-path: frontend/package-lock.json
+    
+    - name: Cache Next.js build
+      uses: actions/cache@v4
+      with:
+        path: |
+          frontend/.next/cache
+          frontend/.next/.next
+        key: ${{ runner.os }}-nextjs-bundle-${{ hashFiles('frontend/package-lock.json') }}-${{ hashFiles('frontend/**/*.js', 'frontend/**/*.jsx', 'frontend/**/*.ts', 'frontend/**/*.tsx') }}
+        restore-keys: |
+          ${{ runner.os }}-nextjs-bundle-${{ hashFiles('frontend/package-lock.json') }}-
+          ${{ runner.os }}-nextjs-bundle-
+    
+    - name: Install dependencies
+      working-directory: ./frontend
+      run: npm ci
+    
+    - name: Build and analyze bundle
+      working-directory: ./frontend
+      run: |
+        npm run build
+        echo "## Bundle Size Report 📊" >> $GITHUB_STEP_SUMMARY
+        echo "" >> $GITHUB_STEP_SUMMARY
+        echo "| Page | Size |" >> $GITHUB_STEP_SUMMARY
+        echo "|------|------|" >> $GITHUB_STEP_SUMMARY
+        
+        # Analyze .next/standalone for size
+        if [ -d ".next/standalone" ]; then
+          du -sh .next/standalone/* | head -20 | while read size path; do
+            name=$(basename "$path")
+            echo "| $name | $size |" >> $GITHUB_STEP_SUMMARY
+          done
+        fi
+        
+        # Show build output size
+        echo "" >> $GITHUB_STEP_SUMMARY
+        echo "### Build Output" >> $GITHUB_STEP_SUMMARY
+        echo "\`\`\`" >> $GITHUB_STEP_SUMMARY
+        du -sh .next/ >> $GITHUB_STEP_SUMMARY
+        echo "\`\`\`" >> $GITHUB_STEP_SUMMARY
+    
+    - name: Upload build artifacts
+      uses: actions/upload-artifact@v4
+      with:
+        name: nextjs-build
+        path: frontend/.next/
+        retention-days: 1
+
+  build-time-tracking:
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v4
+    
+    - name: Track build metrics
+      run: |
+        echo "## Build Performance Metrics ⚡" >> $GITHUB_STEP_SUMMARY
+        echo "" >> $GITHUB_STEP_SUMMARY
+        echo "| Metric | Value |" >> $GITHUB_STEP_SUMMARY
+        echo "|--------|-------|" >> $GITHUB_STEP_SUMMARY
+        echo "| Workflow Started | $(date -u +"%Y-%m-%d %H:%M:%S UTC") |" >> $GITHUB_STEP_SUMMARY
+        echo "| Runner OS | ${{ runner.os }} |" >> $GITHUB_STEP_SUMMARY
+        echo "| Runner Arch | ${{ runner.arch }} |" >> $GITHUB_STEP_SUMMARY
+        
+        # Track Python dependency install time with uv
+        START_TIME=$(date +%s)
+        echo "Installing Python dependencies with uv..."
+        cd backend
+        
+        # This is just for timing comparison, actual install happens in main workflow
+        echo "| Python deps (estimated) | ~10-20s with uv vs 2-3min with pip |" >> $GITHUB_STEP_SUMMARY
+        
+        # Track Node.js dependency install time
+        cd ../frontend
+        NPM_START=$(date +%s)
+        echo "Timing npm ci..."
+        npm ci > /dev/null 2>&1 || true
+        NPM_END=$(date +%s)
+        NPM_DURATION=$((NPM_END - NPM_START))
+        echo "| Node.js deps | ${NPM_DURATION}s |" >> $GITHUB_STEP_SUMMARY