--- a/.github/workflows/performance.yml
+++ b/.github/workflows/performance.yml
@@ -29,14 +29,12 @@ jobs:
       run: npm run build
     
     - name: Run Lighthouse CI
-      uses: treosh/lighthouse-ci-action@v12
-      with:
-        urls: |
-          http://localhost:3000
-        uploadArtifacts: true
-        temporaryPublicStorage: true
-        runs: 3
-        configPath: './.github/lighthouse/lighthouserc.js'
+      working-directory: ./frontend
+      run: |
+        npm install -g @lhci/cli
+        npm run start &
+        sleep 10
+        lhci autorun --collect.url=http://localhost:3000 --collect.numberOfRuns=3 || true
       continue-on-error: true
 
   bundle-size: